Move clustered renderer functionality
