+++ servers/rendering/renderer_rd/forward_clustered/SCsub
@@ -3,6 +3,3 @@
 Import("env")
 
 env.add_source_files(env.servers_sources, "*.cpp")
-
-SConscript("forward_clustered/SCsub")
-SConscript("shaders/SCsub")
